toktok_back/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Pillow==8.1.0
PyJWT==1.7.1
pytz==2020.5
sqlparse==0.4.1
